+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rocketh-deploy
 
+## 0.6.19
+
+### Patch Changes
+
+- fix deterministic deployment when already deployed + showMessage
+- Updated dependencies
+  - rocketh@0.6.17
+
 ## 0.6.18
 
 ### Patch Changes