+++ packages/rocketh-export/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-export
 
+## 0.6.19
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.17
+
 ## 0.6.18
 
 ### Patch Changes