+++ packages/rocketh-signer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-signer
 
+## 0.6.17
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.17
+
 ## 0.6.16
 
 ### Patch Changes