+++ packages/rocketh/CHANGELOG.md
@@ -1,5 +1,11 @@
 # rocketh
 
+## 0.6.17
+
+### Patch Changes
+
+- fix deterministic deployment when already deployed + showMessage
+
 ## 0.6.16
 
 ### Patch Changes